requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==4.2.11
django-allauth==0.62.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
idna==3.7
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.8.0
requests==2.31.0
requests-oauthlib==2.0.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
